requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
